requirements.txt
carla==0.9.16
numpy==2.2.6
opencv-python==4.12.0.88
pygame==2.6.1
PyYAML==6.0.3